Return empty string when calling asString() for $ operators
This was causing exceptions with step file header file name's name when not a string. (Schemawise there is no hard enforcement of filename.name: "name of this exchange structure. It may correspond to the name of the file in a file system or reflect data in this file. There is no strict rule how to use this field.").